Separate mult_test
